t/perf: add tests for ahead-behind
We test both "branch -v" (which should magically get faster
with bitmaps) and the new "ahead/behind", which should beat
out a series of rev-lists.

Here are some timings:

Test                             github              HEAD
----------------------------------------------------------------------------
5312.4: branch -v (walker)       8.92(8.90+0.02)     8.86(8.84+0.03) -0.7%
5312.5: rev-list (walker)        79.53(76.05+2.51)   78.76(75.47+2.36) -1.0%
5312.6: ahead-behind (walker)    7.97(7.94+0.03)     7.92(7.89+0.04) -0.6%
5312.8: branch -v (bitmap)       9.00(8.98+0.02)     0.27(0.26+0.01) -97.0%
5312.9: rev-list (bitmap)        79.10(75.64+2.48)   79.23(75.76+2.48) +0.2%
5312.10: ahead-behind (bitmap)   0.13(0.12+0.01)     0.13(0.11+0.02) +0.0%

Note that the timings for "ahead-behind" on the "github"
branch are bogus; they are picking up the previously built
ahead-behind, which is a bug in the perf system. We can just
ignore them (they should all be "no result").

The interesting things are:

  1. "branch -v" without bitmaps did not get any slower by
     adding the ahead/behind code. This is a good sanity
     check that we did not make things worse in the fallback
     case.

  2. "branch -v" gets way faster with bitmaps (8.8s to
     .27s). Good.

  3. Using ahead-behind is way faster than doing a series of
     rev-lists. (79s versus 0.13s).

We could also try to make the rev-list invocations faster.
They would always suffer from the startup cost (especially
with the pack revidx sorting). But the code would probably
be unnecessarily complex, as we would have to stop
handle_revision_arg from computing the merge base for
"a...b", based on seeing "--count" elsewhere.