ref-filter: use commit_contains
The newly added commit_contains function should do a better
job than our custom depth-first traversal. It should be the
same speed when going close to the roots, but much faster
when all tags are close to the searched-for commit (this
usually isn't the case, but could be if you limit the tags
with a pattern).

It also cleans up some of the more egregious pitfalls of the
original implementation, including an abuse of the
UNINTERESTING and TMP_MARK flags, an utterly confusing
calling convention (it silently caches the bits between
calls, with no checks that our "with_commit" was the same
for each call), and a failure to clean up after itself
(tainting any further traversals).

[I had originally dropped the "with_contains_algo" flag here, as I think
it shoudn't be used anymore. But during a rebase, a lot of this code had
been moved around, so I just left it in. Eventually the old "contains"
algorithm should just go away]

Signed-off-by: Jeff King <peff@peff.net>