diff: use diff-attr hook to determine attributes from content [ci skip]
This is a very rough cut, but it seems to work in a simple
test.

The point is that you could plug in "file" or some other
mechanism to determine which attributes to use. From there,
you can define "diff.foo.binary", or "diff.foo.textconv",
or whatever you like. Attribute names are cached in
git-notes for performance.

Usage is something like:

    # simple and stupid hook example
    cat >.git/hooks/diff-attr <<-\EOF
    #!/bin/sh
    if grep -q "I am a foo file" "$1"; then
	   echo foo
    fi
    EOF
    chmod +x .git/hooks/diff-attr

    # now make an example file
    { echo "I am a foo file"; echo content; } >file
    git add file && git commit -m one
    echo more >>file && git commit -a -m two

    # normal
    git show

    # tweak settings
    git config diff.foo.binary true
    git show

Signed-off-by: Jeff King <peff@peff.net>