xdiff: use st_* helper functions for allocations
The xdiff code is not prepared to handle extremely large
files. Usually this isn't a problem, as we treat large files
as binaries, anyway. But this behavior can be overridden,
and feeding, for example, a 1GB file to xdiff is not pretty.
We may overflow size_t calculations and allocate too-small
buffers.

This patch tries to do the minimal amount to make xdiff
safe, by using git-compat-util's overflow-checking size_t
calculations. Doubtless there are other bugs if you send
such large input to it, but at least we will avoid
overflowing buffers in this case.