xdiff: convert size variables to size_t
libxdiff uses "long" and "int" variables to store some
intermediate size computations, and then feeds the result
malloc. This can cause integer overflow (especially "int",
which is 32-bit even on 64-bit platforms). Usually this just
produces wrong answers, as feeding the same truncated integer
to both malloc and to (for example) memcpy would just result
in a truncated buffer.

But in cases where the size is computed by adding a sequence
of items, the result is fed to malloc, and then the
individual items are copied into the buffer, we can actually
overflow the buffer.

This patch converts these variables to size_t, and uses the
st_add helper to sum them. Note that in fill_conflict_hunk
we do not cover all manipulations of "size"; that's because
the function is called twice: once with a NULL dest to
compute the size, and the second time to fill it (with the
assumption that the caller gave us "size" bytes). We need
only perform the overflow checks on the first round.