log: refactor add_header to drop some magic numbers
We want to chomp newlines off the end of the "value" string.
But because it's const, we must keep its length. This leads
to us having to tweak that length later, to account for
moving the pointer forward.

Since we are about to create a copy of it anyway, let's just
wait and chomp at the end.