pretty: allow formatting names as initials
Initials are shorter and often unique enough in a
per-project setting, so they can be used to give a more
informative version of --oneline.

The 'S' in the placeholder is for "short" (and 's' is
already taken by DATE_SHORT), but obviously that's pretty
arcane.

Possibly there should be more customization of initials,
asking for only 2-letter initials, etc.

Signed-off-by: Jeff King <peff@peff.net>