config.mak: set GIT_PERF_MAKE_OPTS
For years I've just been setting this manually when I run
the perf suite, but that's just silly.

Note that it's not quite foolproof as the suite pulls the
value not out of config.mak, but out of the generated
GIT-BUILD-OPTIONS (so after a change, you need to "make" in
the top-level, even if the workign tree isn't being tested).