ci: use a single worktree
I initially had some notion of testing builds in parallel,
and so put the build for each branch in its own worktree.

But since this is really all about testing whatever is in
HEAD, in practice it runs one at a time (and besides, it
pegs the CPU, so you wouldn't want to run more than one at a
time anyway).  And while multiple worktrees does keep build
artifacts fresh, ccache does a much better job of skipping
unnecessary builds.

On the flipside, there is nothing actually cleaning up the
working trees for old branches, the tmp-ci directory just
grew and grew. Let's get rid of it. It's adding complexity
without benefit.