merged: sort topics with "-t"
By default we just show the topics in alphabetical order.
But the reverse-chronological position is actually much more
interesting, because we'd expect very new topics to be
not-yet-merged, followed by older topics which are merged to
'next', followed by stale or experimental topics that are
not merged at all.

So seeing the "holes" in the clumps of merged topics is a
good way of checking for topics that have fallen through the
cracks.