help: don't look at git_default_config()
There's nothing in the default config callback that should
impact how we show the help. And there's a good reason to
avoid it if we can: running "git help foo" (or "git foo
--help", which triggers that) is a likely thing for somebody
who is trying to figure out how to un-break configuration.
If parsing the config causes us to die, that creates a
chicken and egg problem.