gitconfig: drop push.default=matching
I normally use "current" for most projects. But my pattern
with git.git is to do a mass-push daily (after rebasing),
rather than push individual branches. So long ago I set the
default here to matching, with the thought that I could use
that for the daily push.

But because of the triangular workflow, that didn't actually
work once upon a time (not least, it got the default remote
wrong). So instead, I ended up with the `push` script which
does the right thing.

Since then, Git learned about remote.pushdefault, which
solves part of the problem. So in theory I could get rid of
my script and just use `git push`. But wait; that script
also generates custom refspecs based on which branches need
pushing. And that's harder to do by default (the best I
could do is make sure my branches are in a hierarchy and
push that).

In the meantime, I've come to expect `git push` to just push
the current branch. Even in this daily-push routine, it's
sometimes nice to push up a one-off if I'm planning to share
the link.

So let's just switch to the default of "current", and
continue using `Meta/push` for the daily push.