http: #error on too-old curl
We already fail to build with versions of curl older than
7.19.4. But doing an explicit check with an #error has two
benefits.

One is that it makes it clear to users that the build
failure is intentional, so they don't waste time trying to
debug it.

And two is that it documents our current "too old"
assumption, so that we know whether we need use an #ifdef
when using newer curl features in future patches.