xdl: simplify siphash
We can't just feed our input directly to siphash; we're
simultaneously hashing and finding the end-of-line. My
ancient implementation tried hard to do both of these in a
single pass, but at the cost of complicating the pass. Out
of curiosity, let's see how it behaves if we actually look
over the data in two simpler passes.