Revert "config.mak: use "-g3" for debug builds"
This reverts commit 1e77b8746dc0b192a645dc62decd105e0fa851db.

Wow, this makes the link step _way_ slower. A full build
with the object files ccache'd goes from 4s to 13s. Ouch.
That makes this really annoying to use for things like
bisections or "private -c".

Let's drop it for now. I may want to make it easier to build
with "-g3" on demand, but I'll see if I actually feel the
lack of it.