userdiff: give textconv its own struct
The interface for userdiff_get_textconv() is a little funny.
It takes a "struct userdiff_driver" as input, and returns
that same driver as output, but only if textconv is
configured. This makes it easy to misuse, since you can just
pass the initial userdiff_driver to any function which
expected one to have been passed through the accessor
function. The type system doesn't help us at all.

The reason for this funny design is that the textconv data
is spread across several struct members, and functions like
fill_textconv() need to see all of them.

Let's instead push all of those fields into a "struct
userdiff_textconv" that's embedded inside the
userdiff_driver. Then the get_textconv() accessor can return
just the textconv struct, making it clear that the result
came from the accessor function.