fill_textconv: always check DIFF_FILE_VALID()
When we have an invalid diff_filespec (e.g., "/dev/null" for
an addition or removal) that is not getting a textconv
conversion, we "fill" the buffer by pointing to the empty
string.

But we don't do a similar check when we are doing a
conversion. Instead we let prepare_temp_file() handle it,
which just sets up /dev/null as a tempfile.

This makes the function hard to follow, since it seems like
we're not handling this case (and if we did, that it would
be done inefficiently, running a textconv filter against
/dev/null).

In fact this code path is not triggered at all, since
get_textconv() will not ever return a non-NULL value for an
invalid filespec. But let's hoist the check for
DIFF_FILE_VALID() to the top of the function to make it
clear that we treat invalid files consistently.