diff: convert filespec is_binary flag to enum
We have a tri-state is_binary flag that tells us whether the
file is binary, text, or if we haven't figured it out yet.
In preparation for determining (and caching) other
characteristics of the file, let's switch this to an enum.

There should be no functional changes here, and the enum
values were chosen to be compatible with the original
numeric values.