diff: add utf-{16,32} binary detection
When we examine a diff_filespec for binary-ness, we can now
also determine whether it's utf-16 or utf-32 by looking for
a BOM. Those encodings are still considered binary, but
future patches will add callers which handle them specially.

Note that even though we know the endian-ness of the
encoding from the BOM, we do not include it in the returned
name. This is because "iconv" (which we'll ultimately feed
these names to) does not expect to see a BOM if we tell it
the source is "UTF-16LE". We should feed it "UTF-16" with
the BOM, which it will interpret itself.