pretty: use size_t instead of int for buffer/line sizes
The pretty-print code uses "int" in many places to store
lengths. This is generally fine, as we expect commit objects
and their individual lines to have sane lengths that will
fit in an int. But a malicious commit object could be
gigabytes, causing wrapping or even negative lengths.

This can lead to us producing nonsense results, and even
segfaulting as we access memory outside array bounds. All of
these functions are used for reading the buffer, not
writing. So there should be no possibility of arbitrary code
execution.

Signed-off-by: Jeff King <peff@peff.net>