test: add --retest option
Every once in a while a test fails due to circumstances
outside of the tree itself. It's handy to be able to be able
to trigger a retest.

This probably could stand to be integrate better with the
"-c" option, as "-c --retest" should retest just
compilation. But I don't need it yet, so I'll save that
refactor for another day.