private: allow "fixup" commits for each topic
Traditionally I just lumped any non-conflict fixups on top
of the private branch, and then did a "git cherry-pick
private@{1}" at the end. This is annoying for a few reasons:

  1. You have to remember to do it.

  2. It lumps all of the fixups together. Whereas they
     eventually need to be applied one by one to the
     individual topics (when whatever required the fixup
     finally graduates to master and we rebase).

Doing it individually is even more annoying, because then
you have to pick out the individual fixups and merge them at
the right moment. So let's teach the script to do that.

We'll just carry-forward the fixups from prior rounds, so we
don't need any new storage. We can find them all at the
beginning. This is more efficient, since it requires only a
single invocation of rev-list. But we need to do it anyway
to differentiate them from the non-fixups which require a
force-overwrite.

Ideally we'd store a mapping of branch->hash, but since this
is shell, good luck on any data structure that isn't a
string. Instead, we'll just do a linear search of the fixups
for each topic. This makes the whole thing quadratic in the
number of branches, but in practice there are only a handful
of fixups (and this is better than speculatively doing a
"rev-list --grep" for each topic).

Note the use of a tempfile, too. This is an unfortunate side
effect of shell, as well (we cannot put our loop on the
right-hand side of a pipe, since it would then have no way
to store variables in the outer shell).