private: save $fixup between runs
We only walk the existing "private" branch to find fixups
during the initial invocation. When doing a "--continue",
we'd fail to remember that these fixups need applied. Our
options are basically:

  - traverse again on each --continue

  - save the results in a file

The latter is pretty easy, since we know it's just a single
string with no special parsing required. And it's slightly
more efficient. So let's do that.

It would be simpler still if we just did a single "rev-list
--grep" for each topic, looking for any fixups. But that's
even less efficient, as it's effectively quadratic in the
number of topics.