config.mak: support static $(prefix)
The doc-diff script which I've recently written (and sent
upstream) wants to build and compare the documentation for
two different versions. Having different $(prefix) values
makes that harder (since it looks at the installed result).
We run afoul of this when one of the versions is tagged,
since its prefix is named after the tag (e.g.,
.../v2.17.0/share/man/).

Let's work around this by recognizing the special "omitted"
GIT_VERSION it sets, and using a static prefix.