config.mak: enable DEVELOPER=1
The developer-build support in modern Git covers most of the
CFLAGS I've been setting. So let's use that instead, as it
may grow new ones, too (e.g., -Wformat-security).

There are a couple which it doesn't yet understand, so I'll
leave those here for now as a reminder to suggest them for
inclusion upstream.

Note that we turn this on only in non-strict builds, since
there is no way to turn these into just warnings (because
the DEVELOPER rules are loaded after config.mak is loaded).
This means that in those builds we won't have the warnings
at all. That's probably OK in practice, but an alternative
would be to load config.mak.dev directly here.