drop unused argc parameters
Many functions take an argv/argc pair, but never actually look at argc.
This makes it useless at best (we use the NULL sentinel in argv to find
the end of the array), and misleading at worst (what happens if the argc
count does not match the argv NULL?).

It looks like each of these cases does match, so there's no bug here.
But let's tighten the interfaces to make it harder to get wrong (and to
reduce some -Wunused-parameter complaints).