git-compat-util: add UNUSED keyword
Putting the var name as a parameter makes it hard to
accidentally use the variable marked as "unused".

In theory the compiler could tell us this, but:

  - gcc has no such warning

  - clang has -Wused-but-marked-unused, but it triggers
    false positives with our MAYBE_UNUSED for commit-slab

Note that we have to put it early in git-compat-util.h, because it will
eventually be used in mingw.h (which is itself included in
git-compat-util.h).