private: rewrite to use a "todo" file
When continuing, the "private" script would just go over the
complete list of topics from scratch, skipping those that
had already been merged. This mostly worked, but it made
certain features hacky or unreliable. For instance, if a
topic had a merge fixup, we would forget to apply it if that
topic had a conflict.

Let's try a different approach. When we start a sequence,
we'll generate a list of instructions, similar to how
interactive-rebase works, including those fixup
cherry-picks.  When a command fails, then it becomes trivial
to pick up where we left off.

The diff is hard to read, but the rewrite was pretty
straightforward. It seems to work OK in practice, including
with a failing merge-fixup case.

I suspect that cherry-pick resolution may need to be a
little smarter, but I didn't test it.

In theory this whole thing could be driven by Git's
sequencer machinery, but there isn't really a generic
interface to it yet. It might be possible to just pretend
we're an interactive rebase. However, some of the commands
would probably have to be implemented via arcane "exec"
commands in its todo list. That seems like an area for
future work.