config.mak: turn on verbose test logging
Every once in a while I get a mystery failure that's hard to reproduce,
and it's nice to have a detailed log of what happened.

I did a few timing runs and this appears not to cost very much. Any
runtime changes are within the noise, and it's only about 50MB of logs
per run (which get overwritten on the next run, or deleted on success).