ci: automagically run --status when ci is already running
Most of the time I run this I actually want --status, because I already
kicked off the daemon in the background. This patch lets us detect that
automatically, with two advantages:

 - it's shorter to type

 - I don't accidentally start another daemon, causing them to stomp all
   over each other

The use of `ps` here is pretty horrible. Arguably this should just be
the default mode, and `Meta/ci --daemon` should be the mechanism to
start it.