ci: swap default behavior to --status
As discussed in the previous commit, this is probably a saner default.
And the previous commit's ps trickery is totally broken. It works to
engage "--status" when it should, but it makes it impossible to start
daemon mode, because there is always at least one instance of the script
running (the one that is doing the checking!). I didn't notice because
of course I already had a daemon running while testing it.

This could be fixed in a number of ways, but let's just go with the less
hacky solution. You now have to use "Meta/ci --daemon" to kick off the
daemon.