ci: add verbose status
This just shows the --oneline for each commit alongside its status. This
is unfortunately not possible just by invoking "git log" with --notes,
since the notes are attached to the actual trees. And anyway, it's nice
not to rely on git-test's internal notes structure.

Arguably git-test should provide this, but it doesn't ever look at
anything except the raw sha1s.

The shell here is a little convoluted, but it was a fun exercise to do
it in O(1) processes, rather than looping over the set of commits. The
process substitution does require bash, though. Tempfiles would get
aroudn that (avoid the duplicate "git-test" invocation, too), but are a
whole other world of pain.