private: consider a clean worktree to be resolved
Sometimes a merge or cherry-pick ends up empty; in those cases the
resolution is to do nothing. This lets it work with "--continue".

I ran into this in a real-world case, in which I also managed to get
misaligned with what was in the todo list (but now I don't remember the
details, as it was several days ago). I suspect this may point to a
greater problem in how these resolutions are handled. But lacking a
reproduction, I'll punt on that for now.