gitconfig: enforce strict=yes for git-test builds
For basic runs of the current HEAD, or even ones via rebase (using
Meta/test), config.mak should figure out to run in strict mode. However,
when running Meta/ci, we intentionally move to a detached HEAD, which
triggers the "oops, this might be old, let's be non-strict" settings.

This means that some warnings may be missed, as -Werror is disabled.
Let's turn this on for all git-test runs. There's really no case where I
would actually be visiting old commits with it.