clone: avoid storing URL passwords in config
If you clone a URL with a password, like:

  git clone https://user:pass@example.com/repo.git

we'll write that literal URL, including the password into .git/config.
This can lead to accidentally disclosing it, since .git/config isn't
generally assumed to be a secret. In particular, it's very prone to
accidentally exposing by a webserver:

  1. It's actually _in_ the repo directory, and it's not uncommon for
     people to clone (or copy) the contents to a web-accessible
     directory.

  2. The filesystem permissions are not restrictive. So cloning as user
     "bob" would usually produce a config file readable by user "httpd".

Let's strip the password out before storing it. There are two things to
note:

  - we must (and do) retain the username here. Both as a convenience so
    the user does not have to input it again, but also because
    credential helpers use it as part of the lookup key (which matters
    if you use two different usernames with the same host).

  - since we don't record the password anywhere, a follow-up fetch will
    fail. This is a regression for some workflows, of course, but one
    we'll fix in a future commit. For now we'll warn the user about
    what's happening and add a failing test to show the problem.

XXX This is broken! As written originally, it just replaced
remote->url[0] with the un-stripped "repo". But as noted in cf8f6ce02a
(clone: delay picking a transport until after get_repo_path(),
2023-01-24), it's important to use the remote here, due to f38aa83f9a
(use local cloning if insteadOf makes a local URL, 2014-07-17). I've
replaced this with a comment as a WIP, but of course the tests are going
to fail.

Signed-off-by: Jeff King <peff@peff.net>