ci: use "-f" to create worktree
We usually skip the worktree creation if the directory already exists.
But there's a funny corner case: if we previously had a worktree
registered there, but somehow (e.g., me cleaning cruft) deleted, then
"worktree add" will complain. We need to use "-f" to override that
(which is OK, since we know it is just our temporary dir).

This is basically the same issue that we dealt with upstream in
684e742249 (doc-diff: force worktree add, 2018-08-30).