private: check "-t" script presence for each commit
If we want to test a script if present, then we should be checking for
it at each commit, not wherever the script started from (which might not
even have any relation to what's being merged!).

Let's check at each exec stop whether the script is there and run it
only if so (in fact, we can skip compilation entirely if it isn't).