ci: drop non-verbose status
The "progress bar" line of G/B produced by "--status" is cute, but it's
fairly light on information. The "-v" mode takes the same amount of time
to produce. It might be worse if there are more than a few hundred
commits, but that seems rather unlikely for my use case (besides, I'm
not sure I'd have to patience to run tests on each commit of a 100-patch
series).

Let's drop it entirely, and switch the default status view to verbose.