parse-options: add "intp" to value union
This should give us some type safety when declaring options. This isn't
a complete conversion, so we're actually type-punning a bit behind the
scenes. E.g., OPT_CALLBACK() assigns to value.voidp, but the callback
function might pull it out as value.intp. That's probably OK in
practice, though it's not clear to me if it violates the standard.

The bigger problem, though, is that it gives it works _too_ well. In
particular, we often call OPT_*() macros with pointers to unsigned
integers, rather than signed. Which causes the compiler to complain
(rightfully, as we'd eventually access these via an "int *"!). We could
introduce a matching "uintp" and use it for bit operations, but it seems
that some callers err in the other direction.