pack-objects: trace adjustable window stats for each object
This trace outputs the slot and byte numbers at which we found the delta
candidate that we end up using in the packfile output. If run without
any pack.windowSlotLimit or pack.windowByteLimit, that can give a sense
of what good values are for those config options (at least for this
repository).

Running a full window=200 repack of the torvalds/linux fork network with
GIT_TRACE_DELTA yields a 3GB trace file, from which we can gather some
percentile statistics on the slot numbers:

       p50: 1
       p75: 2
       p90: 8
       p98: 62
       p99: 99
       max: 200

So we usually find our best candidate quite early on. Half the time it's
the first one we try, and 90% of the time it's in the first 8. Of course
there's a tail of cases that require going to 100 or further, but those
are the minority. That explains why limiting to slots=20 worked
reasonably well.

And here's the same run for byte statistics:

       p50: 3462
       p75: 16464
       p90: 97644
       p98: 708408
       p99: 1592577
       max: 265848594

Setting the byte limit at 1m means we'll find the same candidates in
98-99% of cases. But of course the more bytes we look through to find
candidates, the bigger the object, and thus the more we have to gain by
actually finding a delta. So even though our limit excludes less than 2%
of the deltas we'd otherwise found, we still saw our resulting pack size
increase by several percent.

Signed-off-by: Jeff King <peff@peff.net>