pack-objects: make try_delta() result an enum
We already have a tri-state return; in addition to 0 for "yes, we found
a better delta" and 1 for "no we did not", we can return -1 to mean
"no, and do not bother looking further in the window".

Let's turn that into an enum. That makes the existing code a bit more
obvious, and will also let us add further return cases without
sacrificing readability.

Signed-off-by: Jeff King <peff@peff.net>