pack-objects: add window slot limit
The delta window size of pack-objects is a somewhat crude tool for
limiting CPU. For each window slot, we'll try an object against a
candidate source via try_delta(). But that function may do a very small
amount of work (if heuristics tell it that the two objects aren't likely
to delta well) or a large amount (if we actually do a rolling hash over
the candidate bytes to look for similarities).

The difference can be particularly large when delta islands are in
effect, since they will often trigger the "small work" case by
immediately forbidding a delta. To compensate for this, it's useful to
bump the window size when using delta islands, to get better results
while (hopefully) still doing a similar amount of "large work".

But this introduces a new problem: if the island restrictions _don't_
prevent a lot of deltas, then we may end up doing quite a lot of
"large work" cases, spending a lot of CPU but not finding a lot of new
deltas.

Let's introduce a new knob for adjusting the window on the fly: we'll
have a maximum window size as usual, but we'll stop looking through the
window early if we hit more than a certain number of large-window cases.

The new perf test simulates a delta-island situation. Here are the
results on a clone of linux.git:

Test                           origin                  HEAD
-----------------------------------------------------------------------------------
5305.2: window=10              504.82(1317.13+24.38)   491.19(1314.10+25.06) -2.7%
5305.3: pack size                         1.5G                    1.5G -0.0%
5305.4: window=100, slots=10   589.73(3789.62+17.86)   499.10(1415.41+25.53) -15.4%
5305.5: pack size                         1.3G                    1.5G +14.4%

Not amazing. Without this feature (the "origin" column), bumping the
window from 10 to 100 reduces the resulting pack from 1.5G to 1.3G, but
increases the CPU time by almost a factor of 3 (these are multithreaded
repacks, so look at the user CPU, the first number in the parentheses).
The "slots" setting does nothing there, because "origin" doesn't yet
support the feature.

In the second column, we really do set slots=10, and we reclaim most of
that CPU. But we also get a pack that's about the same size as we would
have just by setting window=10 in the first place.

Some real-world tests show better results, though. Here are
comparisons of real from-scratch repacks of the whole GitHub
fork-network of torvalds/linux using delta islands:

  name   window-200   window-20          slots-20
  cpu    66450.04     40487.01 (-39%)    47504.28 (-28%)
  size   1937481275   2695427807 (+39%)  2056879350 (+6%)

The "window-200" column is a repack using window=200, which is what we
normally do. The CPU row is total user CPU, and the size is the size of
a clone of torvalds/linux using the resulting pack.

The third column, slots-20 uses "window=200, slots=20". The resulting
clone is only 6% bigger, but we saved 28% of the CPU during the repack.
Compare that to just setting window=20, we spent a bit more CPU, but our
clone size is much better.  So we've actually accomplished something: a
slider that puts us in the middle of window=200 and window=20.

Here are similar numbers for the git/git network:

  name  window-200  window-20          slots-20
  cpu   82780.5     15301.26 (-81%)    15350.64 (-81%)
  size  146817734   215067328 (+46%)   161205555 (+9%)

Here we do even better: we reclaim almost all of the CPU we could get
with window=20, but our resulting pack is quite a bit smaller.

And finally, on the extreme end, here's a real-world case where the
repository was full of jpegs that don't delta, and had no delta islands
at all:

  name  window-200  window-20          slots-20
  cpu   12577.52    2015.18 (-83%)     1953.42 (-84%)
  size  6211959217  6211960234 (+0%)   6211960234 (+0%)

The pack size doesn't change no matter what we do, but we reclaim the
CPU by using a smaller window. Now obviously you'd do about as well to
just use window=20 here. But the general idea is that we don't know what
the repository looks like ahead of time; by using a larger window=200
but limiting the work with slots=20, we perform reasonably well on
repositories that _do_ have a lot of islands, without paying a penalty
on the ones that don't.

Signed-off-by: Jeff King <peff@peff.net>