pack-objects: add window byte limit
The previous commit introduced a "window slot limit" that would try to
avoid doing too many "expensive" delta attempts within the window. But
"expensive" is relative: looking for deltas in a 1-kilobyte file is much
cheaper than looking for ones in a 1-megabyte file.

Instead of adjusting the delta search based on the number of slots,
let's instead provide the ability to limit based on the number of bytes
examined in each attempt. We can base this on the number of "slow"
slots, since we're really just measuring the effort to walk over the
target object in each slot. One could argue that we should be counting
the sum of the base and the target for each candidate, but the base
effort is all in creating the index, and is amortized over many targets.

Here are numbers for p5305 on a clone of linux.git:

Test                           origin                  HEAD
-----------------------------------------------------------------------------------
5305.2: window=10              504.82(1317.13+24.38)   491.19(1314.10+25.06) -2.7%
5305.3: pack size                         1.5G                    1.5G -0.0%
5305.4: window=100, slots=10   589.73(3789.62+17.86)   499.10(1415.41+25.53) -15.4%
5305.5: pack size                         1.3G                    1.5G +14.4%
5305.6: window=100, bytes=1m   589.73(3789.62+17.86)   572.52(3321.29+19.38) -2.9%
5305.7: pack size                         1.3G                    1.4G +3.6%

Compared to reducing the slot size, using a 1-megabyte limit means we
still end up doing quite a few comparisons (note that we save a much
smaller amount of CPU compared to a true 100-window repack, but our
increase in pack size is much better). So it's arguable which of the two
you'd want to use in this simulated case; they're really just moving the
slider different amounts between the window=10 and window=100 cases.

But let's look at some real-world instances. Here's the GitHub fork
network of torvalds/linux again, this time with some byte limits:

  name  window-200  window-20          slots-20          bytes-10m
  cpu   66450.04    40487.01 (-39%)    47504.28 (-28%)   45344.89 (-31%)
  size  1937481275  2695427807 (+39%)  2056879350 (+6%)  1939474673 (+0%)

We managed to reduce the CPU for the repack by 31% without making any
significant increase in the resulting clone size! It's better in every
metric than the slots-20 run.

Here are benchmarks with a few more byte values:

  name  window-200   bytes-10m         bytes-1m          bytes-500k
  cpu   66450.04     45344.89 (-31%)   46210.42 (-30%)   36448.09 (-45%)
  size  1937481275   1939474673 (+0%)  1998607712 (+3%)  2091254047 (+7%)

We can see the cpu versus size tradeoff as we use smaller limits.
Somewhere between 1m and 10m seems pretty optimal for this particular
repository.

Let's look at the git/git network:

  name  window-200  window-20          slots-20          bytes-1m
  cpu   82780.5     15301.26 (-81%)    15350.64 (-81%)   25172.65 (-69%)
  size  146817734   215067328 (+46%)   161205555 (+9%)   156297226 (+6%)

Here again we spend a bit more CPU than the slots-20 strategy, but we
yield a slightly better packfile. Notice this used a 1-megabyte limit.
Here's the same repository with a few different byte limits:

  name  window-200   bytes-10m         bytes-1m          bytes-500k
  cpu   82780.5      87759.18 (+6%)    25172.65 (-69%)   25435.35 (-69%)
  size  146817734    146848537 (+0%)   156297226 (+6%)   173301406 (+18%)

The 10m limit is clearly too much. The packfile is good, but we spent
even more CPU than without the feature (there's a fair bit of run-to-run
noise, so in reality it's probably about the same). And 500k doesn't buy
us much CPU, but gives us a much worse pack.

And finally, here's that same repository full of jpegs that don't delta:

  name  window-200  window-20         slots-20         bytes-1m          bytes-500k
  cpu   12577.52    2015.18 (-83%)    1953.42 (-84%)   294.03 (-97%)     211.55 (-98%)
  size  6211959217  6211960234 (+0%)  6211960234 (+0%) 6211961945 (+0%)  6211959206 (+0%)

Wow! We get the same pack each time, but now we're using a tenth of the
CPU we'd use even for the small window size. What's going on here?

This is an extreme case: the files are big but don't delta well. So
we're hitting our byte limit almost immediately, and this is behaving
more like a window size of 1 for the blobs, but still finding decent
deltas for smaller blobs and trees. So while the results are impressive,
this repository is atypical. And there's a mirror-image pathological
case: a repository full of large blobs that _do_ delta well, where we
spend very little time looking for deltas and the pack size ends up much
worse.

So this new knob can't magically solve all problems, but it does seem to
improve things for most cases.

Signed-off-by: Jeff King <peff@peff.net>