topics: use rev-parse --parseopt
Commit 92b8e58 (topics: stop using getopts, 2019-12-24) was a bit
over-eager. I've found I actually do type "-vt" all the time.

Going back to getopt is a non-starter because of the whitespace handling
(note that 92b8e58 used the wrong name in the subject). We could switch
to getopts, but it doesn't handle long options. Since we can reasonably
assume that Git is available, let's use its parseopt helper.