rebase: disable editing during "rebase --continue"
Since the move upstream to the merge backend for rebase, Git will open
an editor when continuing a rebase after conflict resolution. For an
automated process like this script (which generally resolves via
rerere), opening an editor would be annoying at best. But because our
stdin is coming from pipe to feed the shell loop, it's disastrous;the
editor barfs because stdin isn't a tty.

Let's just disable this by setting GIT_EDITOR to a noop (upstream may
add a "--no-edit" flag, but there isn't one yet).