private: auto-skip empty cherry-picks
Sometimes a merge-fixup that was required to merge with next graduates
to the individual topic, because the upstream topic made it to master
(and therefore we rebased on top of it, and the problem is now in the
individual topic branch, not in the merge).

That graduation happens manually (and really has to, because it involves
a human figuring out which merge-fixups need to move). But afterwards,
the merge-fixup on next is no longer necessary, but we still try to
cherry-pick it. We probably _could_ use patch-ids to figure this out
(though not reliably, if the form of the fix changed or was folded
into another commit).

But it's easier still to just have cherry-pick notice when the fixup is
no longer necessary, and auto-skip it. We can differentiate this case
from a real cherry-pick conflict by seeing if the index ended up with no
changes.