gitconfig: stop fetching amlog
I had played around with these for a while, but I think it has been at
least a year or two since I actually used them for anything. Let's stop
pulling them to avoid bloating the repository, and it's not that big a
deal to fetch them again if I find a use.