extra: stop summarizing short "PUSH:" output
When many branches are updated (like after running Meta/rebase), the
list of "PUSH:this" and "PUSH:that" becomes uselessly long. So I always
just summarized it as PUSH:all.

But sometimes there's only one or two branches I've been working on, and
those should be pushed. Let's set an arbitrary limit at 3; if we have
more stale branches than that, then we'll summarize it as PUSH:all.